+++ src/EMerchantPay/Libraries/genesis/src/Genesis/API/Request/Financial/Cards/Authorize3D.php
@@ -27,6 +27,7 @@ namespace Genesis\API\Request\Financial\Cards;
 
 use Genesis\API\Traits\Request\Financial\Business\BusinessAttributes;
 use Genesis\API\Traits\Request\Financial\Cards\Recurring\ManagedRecurringAttributes;
+use Genesis\API\Traits\Request\Financial\Cards\Recurring\RecurringCategoryAttributes;
 use Genesis\API\Traits\Request\Financial\Cards\Recurring\RecurringTypeAttributes;
 use Genesis\API\Traits\Request\Financial\CryptoAttributes;
 use Genesis\API\Traits\Request\Financial\FxRateAttributes;
@@ -59,7 +60,7 @@ class Authorize3D extends \Genesis\API\Request\Base\Financial\Cards\CreditCard
     use GamingAttributes, MotoAttributes, NotificationAttributes, AsyncAttributes, AddressInfoAttributes,
         MpiAttributes, RiskAttributes, DescriptorAttributes, PreauthorizationAttributes, TravelDataAttributes,
         ScaAttributes, FxRateAttributes, CryptoAttributes, BusinessAttributes, RestrictedSetter,
-        AllThreedsV2Attributes, RecurringTypeAttributes, ManagedRecurringAttributes;
+        AllThreedsV2Attributes, RecurringTypeAttributes, ManagedRecurringAttributes, RecurringCategoryAttributes;
 
     /**
      * Returns the Request transaction type
@@ -71,13 +72,22 @@ class Authorize3D extends \Genesis\API\Request\Base\Financial\Cards\CreditCard
     }
 
     /**
-     * Transaction Request with zero amount is allowed
+     * Return the required parameters keys which values could evaluate as empty
+     * Example value:
+     * array(
+     *     'class_property' => 'request_structure_key'
+     * )
      *
-     * @return bool
+     * @return array
      */
-    protected function allowedZeroAmount()
+    protected function allowedEmptyNotNullFields()
     {
-        return true;
+        return array_merge(
+            array(
+                'amount' => static::REQUEST_KEY_AMOUNT
+            ),
+            $this->getAllowedFieldsZeroValues()
+        );
     }
 
     /**
@@ -158,7 +168,8 @@ class Authorize3D extends \Genesis\API\Request\Base\Financial\Cards\CreditCard
                 'business_attributes'       => $this->getBusinessAttributesStructure(),
                 'threeds_v2_params'         => $this->getThreedsV2ParamsStructure(),
                 'recurring_type'            => $this->getRecurringType(),
-                'managed_recurring'         => $this->getManagedRecurringAttributesStructure()
+                'managed_recurring'         => $this->getManagedRecurringAttributesStructure(),
+                'recurring_category'        => $this->recurring_category
             ],
             $this->getScaAttributesStructure()
         );