+++ src/EMerchantPay/Libraries/genesis/src/Genesis/API/Traits/Request/Financial/Cards/Recurring/RecurringTypeAttributes.php
@@ -106,7 +106,7 @@ trait RecurringTypeAttributes
         return [
             'recurring_type' => [
                 Types::MANAGED => [
-                    'managed_recurring_period'
+                    'managed_recurring_mode'
                 ]
             ]
         ];